requirements.txt
Django==2.2
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2023.3
sqlparse==0.4.4
